+++ Count my breaths/breathCounter/components/RecordView.swift
