+++ Count my breaths/breathCounter/pages/CountView.swift
