+++ Count my breaths/breathCounter/pages/GraphView.swift
