+++ Count my breaths/breathCounter/pages/HistoryView.swift
