+++ Count my breaths/breathCounter/pages/HowToView.swift
