+++ Count my breaths/breathCounter/pages/MailView.swift
@@ -1,0 +1,21 @@
+//
+//  MailView.swift
+//  Count my breaths
+//
+//  Created by Lucas Desouza on 10/20/20.
+//  Copyright © 2020 Lucas Desouza. All rights reserved.
+//
+
+import SwiftUI
+
+struct MailView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct MailView_Previews: PreviewProvider {
+    static var previews: some View {
+        MailView()
+    }
+}
