+++ Count my breaths/breathCounter/pages/StatsView.swift
