--- Count my breaths/components/MailView.swift
@@ -1,80 +1,0 @@
-//
-// Created by Lucas Desouza on 5/23/20.
-// Copyright (c) 2020 Lucas Desouza. All rights reserved.
-//
-
-import MessageUI
-import QuickComponents
-import SwiftUI
-
-struct MailButtonView: View {
-
-    @State var result: Result<MFMailComposeResult, Error>? = nil
-    @State var isShowingMailView = false
-    @State var csvData: Data? = nil
-
-    var body: some View {
-        VStack {
-            if MFMailComposeViewController.canSendMail() {
-                Button("Email records") {
-                    self.isShowingMailView.toggle()
-                }.frame(width:150, height:50).buttonStyle(PrimaryButton(variant: .empty))
-            }
-        }.sheet(isPresented: $isShowingMailView) {
-            MailView(isShowing: self.$isShowingMailView, result: self.$result, csv: self.$csvData)
-        }
-    }
-}
-
-struct MailView: UIViewControllerRepresentable {
-
-    @Binding var isShowing: Bool
-    @Binding var result: Result<MFMailComposeResult, Error>?
-    @Binding var csv: Data?
-
-    class Coordinator: NSObject, MFMailComposeViewControllerDelegate {
-
-        @Binding var isShowing: Bool
-        @Binding var result: Result<MFMailComposeResult, Error>?
-        @Binding var csv: Data?
-
-        init(isShowing: Binding<Bool>,
-             result: Binding<Result<MFMailComposeResult, Error>?>, csv: Binding<Data?>) {
-            _isShowing = isShowing
-            _result = result
-            _csv = csv
-        }
-
-        func mailComposeController(_ controller: MFMailComposeViewController,
-                                   didFinishWith result: MFMailComposeResult,
-                                   error: Error?) {
-            defer {
-                isShowing = false
-            }
-            guard error == nil else {
-                self.result = .failure(error!)
-                return
-            }
-            self.result = .success(result)
-        }
-    }
-
-    func makeCoordinator() -> Coordinator {
-        return Coordinator(isShowing: $isShowing,
-                result: $result, csv: $csv)
-    }
-
-    func makeUIViewController(context: UIViewControllerRepresentableContext<MailView>) -> MFMailComposeViewController {
-        let vc = MFMailComposeViewController()
-        vc.mailComposeDelegate = context.coordinator
-        if self.csv != nil {
-            vc.addAttachmentData(self.csv!, mimeType: "text/csv", fileName: "breaths-per-minute.csv")
-        }
-        return vc
-    }
-
-    func updateUIViewController(_ uiViewController: MFMailComposeViewController,
-                                context: UIViewControllerRepresentableContext<MailView>) {
-
-    }
-}
