--- Count my breaths/views/AccountView.swift
@@ -1,24 +1,0 @@
-//
-//  AccountView.swift
-//  Count my breaths
-//
-//  Created by Lucas Desouza on 5/12/20.
-//  Copyright © 2020 Lucas Desouza. All rights reserved.
-//
-
-import SwiftUI
-
-struct AccountView: View {
-    var body: some View {
-        VStack {
-            HeaderView(title:"Account", subTitle: "Comming soon...")
-            Spacer()
-        }
-    }
-}
-
-struct AccountView_Previews: PreviewProvider {
-    static var previews: some View {
-        AccountView()
-    }
-}
